django-api/pyproject.toml
[tool.poetry]
name = "django-api"
version = "0.1.0"
description = ""
authors = ["Nicholas Ferrara <nferrara100@gmail.com>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
requests = "^2.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
markdown = "^3.4.1"
mypy = "^0.991"
bandit = "^1.7.4"
flake8 = "^5.0.4"
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
